[Refactor]: 로컬, 운영 레포지토리 분리
