+++ back/src/main/java/com/back/domain/post/entity/Post.java
@@ -15,6 +15,7 @@ import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.service.spi.ServiceException;
 import org.hibernate.type.SqlTypes;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.annotation.LastModifiedDate;
 
 import java.time.LocalDateTime;
@@ -30,7 +31,6 @@ import java.util.List;
                         columnList = "category, created_date DESC"),
                 @Index(name = "idx_post_user_created",
                         columnList = "user_id, created_date DESC"),
-//                @Index(name = "idx_post_title", columnList = "title")
         })
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
@@ -77,14 +77,13 @@ public class Post extends BaseEntity {
     private Scenario scenario;
 
     /**
-     * PostgreSQL tsvector 타입 컬럼
+     * 마이그레이션 alter문으로 postgreSQL tsvector 타입 컬럼 지정
      * - 트리거에 의해 자동으로 업데이트됨
      * - 직접 값을 설정할 필요 없음
      * - GIN 인덱스로 빠른 검색 지원
      */
     @Column(
             name = "search_vector",
-            columnDefinition = "tsvector",
             insertable = false,
             updatable = false
     )
