+++ back/src/main/java/com/back/domain/post/repository/PostRepository.java
@@ -16,7 +16,7 @@ import java.util.Optional;
  * 게시글 엔티티에 대한 데이터베이스 접근을 담당하는 JpaRepository.
  */
 @Repository
-public interface PostRepository extends JpaRepository<Post, Long>, PostRepositoryCustom {
+public interface PostRepository extends JpaRepository<Post, Long> {
 //    @Lock(LockModeType.PESSIMISTIC_WRITE)
 //    @Query("SELECT p FROM Post p WHERE p.id = :postId")
 //    Optional<Post> findByIdWithLock(@Param("postId") Long postId);
